backend/requirements.txt
fastapi>=0.121.3
sqlmodel>=0.0.27
uvicorn>=0.18.3
python-dotenv>=1.0.0
openai>=1.0.0
aiofiles>=23.0.0
